Fix some broken URLs and missing translation
Several URLs are broken in FR page. Care, I don't know how the doc is built hence, how the corrected links will be rendered...
Curiously, some sentences are not translated in FR in this source file, contrary to the online docs. I've missed smthg... but here is the correct FR translation.